compiling plastic utest. It does not read properly the thickness of the problem

+++ src/nb/pde_bot/finite_element/solid_mechanics/static_plasticity2D.c
@@ -21,7 +21,7 @@
 #include "nb/pde_bot/finite_element/gaussp_to_nodes.h"
 #include "nb/pde_bot/finite_element/solid_mechanics/static_elasticity2D.h"
 
-#include "set_bconditions.h"
+#include "nb/pde_bot/finite_element/solid_mechanics/set_bconditions.h"
 #include "pipeline.h"
 #include "nb/pde_bot/common_solid_mechanics/analysis2D.h"
 #include "nb/pde_bot/finite_element/solid_mechanics/static_plasticity2D.h"
@@ -41,7 +41,7 @@ int fem_compute_plastic_2D_Solid_Mechanics
 			 double *total_strain, /* Output*/
 			 double *stress,
 			 double *displacement, /* Output, just the last computed plastic displacement */
-			 double N_force_steps,
+			 uint32_t N_force_steps,
 			 double yield_stress,
 			 double accepted_tol)
 {
